[refactor] 불필요한 객체 삭제 및 통합
- 기존의 User 객체를 사용하던 CustomUserDetails를 User Entity와 통합하였음
- 이후 CustomUserDetails 삭제
