--- src/main/java/org/board/exercise_board/user/Security/CustomUserDetails.java
@@ -1,52 +1,0 @@
-package org.board.exercise_board.user.Security;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import lombok.RequiredArgsConstructor;
-import org.board.exercise_board.user.domain.model.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-@RequiredArgsConstructor
-public class CustomUserDetails implements UserDetails {
-
-  private final User user;
-
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-    Collection<GrantedAuthority> authorities = new ArrayList<>();
-    authorities.add(new SimpleGrantedAuthority(user.getRole().toString()));
-    return authorities;
-  }
-
-  @Override
-  public String getPassword() {
-    return user.getPassword();
-  }
-
-  @Override
-  public String getUsername() {
-    return user.getLoginId();
-  }
-
-  @Override
-  public boolean isAccountNonExpired() {
-    return true;
-  }
-
-  @Override
-  public boolean isAccountNonLocked() {
-    return true;
-  }
-
-  @Override
-  public boolean isCredentialsNonExpired() {
-    return true;
-  }
-
-  @Override
-  public boolean isEnabled() {
-    return true;
-  }
-}
